requirements.txt
torch==2.3.1
sentence-transformers==3.0.1
annoy==1.17.3
requests==2.32.3
beautifulsoup4==4.12.3
html2text==2024.2.26
tqdm==4.66.4
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
pyyaml==6.0.1
